// typescript-projects/utility-backups/src20250516BeforeErrorHandlingChange/interface/types.ts
export type Flatten<T> = T extends Record<string, unknown>
  ? UnionToIntersection<
      {
        [K in keyof DeepRequired<T>]: K extends string
          ? T[K] extends Record<string, unknown>
            ? {
                [P in `${K}.${keyof Flatten<T[K]> & string}`]: Flatten<
                  T[K]
                >[P extends `${K}.${infer R}` ? R : never];
              } & { [P in K]: T[K] }
            : { [P in K]: T[K] }
          : never;
      }[keyof T]
    >
  : never;

export type UnionToIntersection<T> = (
  T extends unknown ? (x: T) => void : never
) extends (x: infer R) => void
  ? R
  : never;

/** Recursively make all properties of a map/object required */
export type DeepRequired<T> = {
  [K in keyof T]-?: Exclude<T[K], undefined> extends Record<string, unknown>
    ? DeepRequired<T[K]>
    : T[K];
};
// ...
export type TAppPaymentData = {
  /** Unique identifier for the customer, at the merchant */
  customerId: string;
  /** Producer of value. E.g. mtn, dstv, betking, ikeja-electric */
  merchantId: string;
  /** The unique identifier we generate and send to provider...per transaction */
  appOrderId: string;
  /** ID for a bundled, fixed-price products. E.g. ID for internet and cable tv product bundles/plans/bouquet */
  productId?: string;
  /** Amount for a variable-price product like airtime */
  amount?: number;
} & ( // When productId is provided, there's no need for an amount, and vice versa
  | { productId: string; amount?: never }
  | { amount: number; productId?: never }
);

/** Data expected to come into the payment-processing endpoint service class */
export type TProcessData = {
  paymentData: TAppPaymentData;
  uid: string;
  appTxHash: string;
  nairaAdminDocPath: string;
  txPath: string;
  provider: BaseProvider;
  providerId: ProviderId;
};
// ...
export type TStatusDocsAccessData = {
  uid: string;
  /** hash of naira withdrawal tx, which also serves as tx doc ID */
  appTxHash: string;
  /** full path to the related naira_admin doc */
  nairaAdminDocPath: string;
};
// ...
export type TNairaMetadataUtilityWriteData = {
  utilityStatus?: AppTxStatus;
  /** ID of the utility payment tx we generate and supply to the provider */
  appOrderId?: string;
  /** ID of the utility payment tx the provider generates and sends back to us */
  providerOrderId?: string;
  txTimestamp?: number;
  providerId?: ProviderId;
  txPath?: string;
};